+++ crates/swc_ecma_transforms_compat/tests/static-blocks/class-properties/new-target/output.js
@@ -1,12 +1,13 @@
 class Base {
-    constructor() {
-        this.Foo = (function () {
-            class _class {}
+    constructor(){
+        this.Foo = (function() {
+            class _class {
+            }
             var __ = {
                 writable: true,
-                value: (() => {
-                    _class.foo = new.target;
-                })(),
+                value: (()=>{
+                    _class.foo = void 0;
+                })()
             };
             return _class;
         })();